requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
tqdm==4.66.4
langdetect==1.0.9
transformers==4.44.2
accelerate==0.33.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
networkx==3.3
pyarrow==17.0.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
langchain==0.2.12
langchain-community==0.2.11
neo4j==5.19.0
python-dotenv==1.0.1
ijson==3.2.3
regex==2024.5.15
